+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -39,7 +39,7 @@ export { default as updatePaymentRequest } from './updatePaymentRequest';
 export { default as updatePaymentRequestJobRunner } from './updatePaymentRequestJobRunner';
 export { default as updatePaymentRequestStatus } from './updatePaymentRequestStatus';
 export { default as updateReweighWeight } from './updateReweighWeight';
-export { default as updateServiceItemStatus } from './updateServiceItemStatus';
+export { default as updateServiceItemStatus } from './UpdateServiceItemStatus/updateServiceItemStatus';
 export { default as updateMTOReviewedBillableWeightsAt } from './updateMTOReviewedBillableWeightsAt';
 export { default as uploadAmendedOrders } from './UploadAmendedOrders/uploadAmendedOrders';
 export { default as updateMTOShipmentStatus } from './updateMTOShipmentStatus';
